Add new `ape-pie` client
