+++ .github/workflows/ci.yml
@@ -20,14 +20,10 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python: ['3.7', '3.8', '3.9', '3.10']
+        python: ['3.10', '3.11', '3.12']
         django: ['3.2', '4.1', '4.2']
         zdsclient: ['1.0', '2.0']
         exclude:
-          - python: '3.7'
-            django: '4.1'
-          - python: '3.7'
-            django: '4.2'
           - python: '3.11'
             django: '3.2'
 
@@ -48,7 +44,7 @@ jobs:
           - 6379:6379
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python }}
@@ -74,10 +70,10 @@ jobs:
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
         with:
-          python-version: '3.9'
+          python-version: '3.10'
 
       - name: Build sdist and wheel
         run: |
@@ -85,7 +81,7 @@ jobs:
           python setup.py sdist bdist_wheel
 
       - name: Publish a Python distribution to PyPI
-        uses: pypa/gh-action-pypi-publish@v1.5.1
+        uses: pypa/gh-action-pypi-publish@v1
         with:
           user: __token__
           password: ${{ secrets.PYPI_TOKEN }}
