+++ .github/workflows/code_quality.yml
@@ -27,10 +27,10 @@ jobs:
           - docs
           # - flake8
     steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
         with:
-          python-version: '3.7'
+          python-version: '3.10'
       - name: Install dependencies
         run: pip install tox
       - run: tox
