+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =
     py37-django32-zdsclient{10,20}
-    py{38,39,310}-django{32,41,42}-zdsclient{10,20}
+    py{310,311,312}-django{32,41,42}-zdsclient{10,20}
     py311-django{41,42}-zdsclient{10,20}
     isort
     black
@@ -11,10 +11,9 @@ skip_missing_interpreters = true
 
 [gh-actions]
 python =
-    3.7: py37
-    3.8: py38
-    3.9: py39
     3.10: py310
+    3.11: py311
+    3.12: py312
 
 [gh-actions:env]
 DJANGO =
